c-shquote: implement library
The quote and unquote functions take an input string and a pointer to
an output string. On success, the output pointer is updated to
point to the end of the successfully written string, on failure,
the pointer remains unchanged, but the underlying buffer may have
been written to and its contents is undefined.

The argument parser also takes a pointer to the input string, as
well as the output and walks it in the same way. When a token has
been parsed 0 is returned, and when there are no more tokens
C_SHQUOTE_E_EOF is returned. Tokens can be zero-length.

The semantics of these functions are as defined by glib, which is
compatible with /bin/sh, with some restrictions.

Signed-off-by: Tom Gundersen <teg@jklm.no>