c-shquote: add explicit assertion
In c_shquote_parse_next(), the last fallback fast-forwards to the next
escape sequence. Add an explicit assertion to make sure the
fast-forward does not contain more characters than out switch()
statement before.

While at it, order the special characters in the fast-foward the same
as in our switch-statement to make it easier to review.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>