c-shquote: rework 'quoted' to 'got_output'
Use a more descriptive boolean state that describes whether the current
state in c_shquote_parse_next() has produced output to be returned, or
not. This does not imply that the output is non-empty, but it rather
implies that we must consider the output to be returned to the caller.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>